requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-telegram-bot[webhooks]>=20.7
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
httpx>=0.27
